agctl: move trace and config command logic under proxy subcommand
The canonical `agctl proxy trace` and `agctl proxy config` commands
previously imported their definitions from the top-level
`controller/pkg/cli/{trace,config}` packages, making the deprecated
top-level aliases the source of truth.

Invert this so the command definitions and underlying logic live under
`controller/pkg/cli/proxy/{trace,config}`, and reduce the top-level
`trace`/`config` packages to thin re-export shims used only by the
deprecated `agctl trace` / `agctl config` aliases.

No behavior change: `agctl proxy trace|config` work as before and the
deprecated top-level aliases still print their deprecation notice.

Fixes #2169
