+++ controller/pkg/cli/proxy/config/table.go
