+++ controller/pkg/cli/proxy/trace/trace.go
