+++ controller/pkg/cli/proxy/trace/trace_test.go
